Refactor lambda alarm code
