+++ src/constructs/cloudwatch/lambda-alarms.ts
@@ -4,20 +4,14 @@ import type { GuLambdaFunction } from "../lambda";
 import type { GuAlarmProps } from "./alarm";
 import { GuAlarm } from "./alarm";
 
-export type LambdaMonitoring = NoMonitoring | ErrorPercentageMonitoring;
-
-export interface NoMonitoring {
-  noMonitoring: true;
-}
-
-export interface ErrorPercentageMonitoring
+export interface LambdaErrorPercentageMonitoring
   extends Omit<GuAlarmProps, "metric" | "threshold" | "comparisonOperator" | "evaluationPeriods"> {
   toleratedErrorPercentage: number;
   numberOfFiveMinutePeriodsToEvaluate?: number;
   noMonitoring?: false;
 }
 
-interface GuLambdaAlarmProps extends ErrorPercentageMonitoring {
+interface GuLambdaAlarmProps extends LambdaErrorPercentageMonitoring {
   lambda: GuLambdaFunction;
 }
