+++ src/constructs/cloudwatch/no-monitoring.ts
@@ -1,0 +1,3 @@
+export interface NoMonitoring {
+  noMonitoring: true;
+}
