+++ src/patterns/scheduled-lambda.ts
@@ -1,13 +1,13 @@
 import type { Schedule } from "@aws-cdk/aws-events";
-import type { LambdaMonitoring } from "../constructs/cloudwatch/lambda-alarms";
-import { GuLambdaErrorPercentageAlarm } from "../constructs/cloudwatch/lambda-alarms";
+import type { LambdaErrorPercentageMonitoring } from "../constructs/cloudwatch/lambda-alarms";
+import type { NoMonitoring } from "../constructs/cloudwatch/no-monitoring";
 import type { GuStack } from "../constructs/core";
 import { GuLambdaFunction } from "../constructs/lambda";
 import type { GuFunctionProps } from "../constructs/lambda";
 
 interface GuScheduledLambdaProps extends Omit<GuFunctionProps, "rules" | "apis"> {
   schedule: Schedule;
-  monitoringConfiguration: LambdaMonitoring;
+  monitoringConfiguration: NoMonitoring | LambdaErrorPercentageMonitoring;
 }
 
 export class GuScheduledLambda extends GuLambdaFunction {
@@ -15,13 +15,8 @@ export class GuScheduledLambda extends GuLambdaFunction {
     const lambdaProps: GuFunctionProps = {
       ...props,
       rules: [{ schedule: props.schedule }],
+      errorPercentageMonitoring: props.monitoringConfiguration.noMonitoring ? undefined : props.monitoringConfiguration,
     };
     super(scope, id, lambdaProps);
-    if (!props.monitoringConfiguration.noMonitoring) {
-      new GuLambdaErrorPercentageAlarm(scope, "error-percentage-alarm-for-scheduled-lambda", {
-        ...props.monitoringConfiguration,
-        lambda: this,
-      });
-    }
   }
 }
